# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Projet API",
    "description": "Proof-of-work hiring infrastructure.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/p/{handle}": {
      "get": {
        "tags": [
          "public profile"
        ],
        "summary": "Get Public Profile",
        "operationId": "get_public_profile_p__handle__get",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/{key}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Serve File",
        "description": "Section 8 \u2014 CVs and snapshots are not publicly addressable.\n\nThe signature is scoped to the exact key and expires, so a leaked URL\nstops working and cannot be edited to reach a different file.",
        "operationId": "serve_file_files__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicAttestedSkill": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "attesters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attesters"
          },
          "programme_count": {
            "type": "integer",
            "title": "Programme Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "attesters",
          "programme_count"
        ],
        "title": "PublicAttestedSkill"
      },
      "PublicCredential": {
        "properties": {
          "company": {
            "type": "string",
            "title": "Company"
          },
          "programme": {
            "type": "string",
            "title": "Programme"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          },
          "attesters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Attesters"
          },
          "start_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "type": "object",
        "required": [
          "company",
          "programme",
          "skills",
          "attesters",
          "start_at",
          "ended_at"
        ],
        "title": "PublicCredential"
      },
...
      "PublicProfile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "handle": {
            "type": "string",
            "title": "Handle"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/PublicAttestedSkill"
            },
            "type": "array",
            "title": "Skills"
          },
          "attester_count": {
            "type": "integer",
            "title": "Attester Count"
          },
          "programme_count": {
            "type": "integer",
            "title": "Programme Count"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/PublicProjectEntry"
            },
            "type": "array",
            "title": "Projects"
          },
          "credentials": {
            "items": {
              "$ref": "#/components/schemas/PublicCredential"
            },
            "type": "array",
            "title": "Credentials"
          },
          "testimonials": {
            "items": {
              "$ref": "#/components/schemas/PublicTestimonial"
            },
            "type": "array",
            "title": "Testimonials"
          },
          "verified_project_count": {
            "type": "integer",
            "title": "Verified Project Count"
          },
          "self_declared_project_count": {
            "type": "integer",
            "title": "Self Declared Project Count"
          },
          "programmes_completed": {
            "type": "integer",
            "title": "Programmes Completed"
          }
        },
        "type": "object",
        "required": [
          "name",
          "handle",
          "headline",
          "bio",
          "location",
          "skills",
          "attester_count",
          "programme_count",
          "projects",
          "credentials",
          "testimonials",
          "verified_project_count",
          "self_declared_project_count",
          "programmes_completed"
        ],
        "title": "PublicProfile"
      },
      "PublicProjectEntry": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "associated_experience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Associated Experience"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "ongoing": {
            "type": "boolean",
            "title": "Ongoing"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/PublicProjectLink"
            },
            "type": "array",
            "title": "Links"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills"
          }
        },
        "type": "object",
        "required": [
          "verified",
          "title",
          "associated_experience",
          "started_at",
          "ended_at",
          "ongoing",
          "description",
          "links",
          "skills"
        ],
        "title": "PublicProjectEntry"
      },
      "PublicProjectLink": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "url",
          "filename",
          "label"
        ],
        "title": "PublicProjectLink",
        "description": "A URL, or a signed expiring link to an uploaded file."
      },
      "PublicTestimonial": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          },
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Name"
          },
          "author_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "programme": {
            "type": "string",
            "title": "Programme"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "body",
          "author_name",
          "author_title",
          "company",
          "programme",
          "pdf_url",
          "published_at"
        ],
        "title": "PublicTestimonial"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}